requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ── Feature 2: RAG pipeline ──
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
numpy>=1.24.0
torch>=2.0.0

# ── Feature 3: FastAPI backend ──
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pyyaml>=6.0
python-dotenv>=1.0.0
